[PlayCode_5_15] Use force transition to solve the declaration problem.

+++ Chapter5/5_15/PlayCode/pound.c
@@ -9,7 +9,7 @@ int main(void)
 
 	pound(iTest);
 	pound(cTest);
-	pound(fTest);
+	pound((int)fTest);
 
 	return 0;
 }
